requirements.txt
torch 
torchvision
numpy 
matplotlib
pillow 
pyyaml
scikit-learn
timm 